+++ src/main/java/ru/netcracker/registration/service/SpotService.java
@@ -7,10 +7,10 @@ import ru.netcracker.registration.model.Spot;
 import java.util.List;
 
 public interface SpotService {
-    Spot getById(Integer id);
+    Spot getById(Long id);
     List<Spot> getByName(String name);
     List<Spot> getAll();
-    boolean delete(Integer id);
+    boolean delete(Long id);
     boolean delete(Spot spot);
     Spot save(Spot spot);
     List<SpotDTO> getAllInRange(double lat, double lng, double range);
